Normalized heat flow report

+++ gui/src/main/java/io/github/therealmone/fireres/gui/controller/heat/flow/HeatFlowParamsController.java
@@ -26,7 +26,7 @@ public class HeatFlowParamsController extends AbstractController implements Heat
     private Spinner<Integer> sensorSpinner;
 
     @FXML
-    private Spinner<Integer> heatFlowBoundSpinner;
+    private Spinner<Double> heatFlowBoundSpinner;
 
     @Inject
     private ResetSettingsService resetSettingsService;
